refactor: moved jobs bounded context to its own microservice
