+++ api-jobs/src/main/java/ga/wawupc/api/jobs/AppConfig.java
@@ -1,0 +1,13 @@
+package ga.wawupc.api.jobs;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
+
+@Configuration
+public class AppConfig {
+  @Bean
+  public LocalValidatorFactoryBean validator() {
+    return new LocalValidatorFactoryBean();
+  }
+}
