+++ api-jobs/src/main/java/ga/wawupc/api/jobs/domain/model/entity/Offer.java
