+++ api-jobs/src/main/java/ga/wawupc/api/jobs/mapping/MappingConfiguration.java
