+++ api-jobs/src/main/java/ga/wawupc/api/jobs/resource/OfferRequest.java
