+++ api-jobs/src/main/resources/application-h2.properties
@@ -1,0 +1,12 @@
+# H2 In-Memory Database
+spring.h2.console.enabled = true
+spring.h2.console.path = /_/h2-console
+# Datasource Configuration
+spring.datasource.url = jdbc:h2:mem:;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;MODE=MySQL;DATABASE_TO_UPPER=FALSE
+spring.datasource.username = sa
+spring.datasource.password = admin
+# JPA Configuration
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.H2Dialect
+# Migrations
+spring.jpa.hibernate.ddl-auto = update
+spring.flyway.enabled = false
