+++ api-jobs/src/main/resources/application-h2persistent.properties
@@ -1,0 +1,2 @@
+# Datasource Configuration
+spring.datasource.url = jdbc:h2:file:./.cache/databases/jobs;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;MODE=MySQL;DATABASE_TO_UPPER=FALSE
