+++ api-jobs/src/main/resources/application-mysql.example.properties
@@ -1,0 +1,6 @@
+# Datasource Configuration
+spring.datasource.url = jdbc:mysql://localhost:3306/waw_jobs
+spring.datasource.username = {YOUR_USERNAME}
+spring.datasource.password = {YOUR_PASSWORD}
+# JPA Configuration
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQLDialect
