+++ api-jobs/src/main/resources/application-mysqlenv.properties
@@ -1,0 +1,6 @@
+# Datasource Configuration
+spring.datasource.url = jdbc:mysql://${MYSQL_HOST:localhost}:${MYSQL_PORT:3306}/${MYSQL_DB_NAME:waw_jobs}
+spring.datasource.username = ${MYSQL_USERNAME}
+spring.datasource.password = ${MYSQL_PASSWORD}
+# JPA Configuration
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQLDialect
