+++ api-jobs/src/main/resources/application-prod.properties
@@ -1,0 +1,11 @@
+# Hibernate SQL Logging
+logging.level.org.hibernate.SQL = info
+logging.level.org.hibernate.type = info
+# Migrations
+spring.jpa.hibernate.ddl-auto = none
+spring.flyway.enabled = true
+# H2-specific configurations
+# H2 shouldn't be in use in production, but it doesn't hurt to set this
+spring.h2.console.enabled = false
+# Dev Tools
+spring.devtools.add-properties = false
