+++ api-jobs/src/main/resources/application.properties
@@ -1,0 +1,23 @@
+# Datasource Configuration
+spring.cloud.gcp.core.enabled = false
+spring.cloud.gcp.sql.enabled = false
+spring.profiles.active = h2, h2persistent
+# Hibernate Configuration
+spring.jpa.hibernate.ddl-auto = update
+spring.flyway.enabled = false
+# Hibernate SQL Logging
+logging.level.org.hibernate.SQL = debug
+logging.level.org.hibernate.type = trace
+# Error Stacktrace
+server.error.include-stacktrace = never
+server.error.include-message = always
+# Configure Port
+server.port = ${PORT:8080}
+# Allow Bean dependencies
+spring.main.allow-circular-references = true
+# Swagger UI
+springdoc.swagger-ui.operations-sorter = alpha
+# Explicitly set open in view to supress warning
+spring.jpa.open-in-view = true
+# Spring Boot Actuator
+management.endpoints.web.base-path = /_/actuator
