+++ api-jobs/src/main/resources/db/migration/V20220614014354__CreateOffers.sql
